requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Login==0.6.2
python-dotenv==0.19.0
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.46